Convert fetching commands to standard management commands
